// src/app/core/enums/numerotation.enum.ts
/**
 * enum describe numerotation types
 */
export enum NumerationType {
    FACTURE = 1,
    AVOIR = 2,
    USER = 3,
    CLIENT_PROFESSIONNEL = 4,
    CLIENT_PARTICULIER = 5,
    CLIENT_OBLIGES = 6,
    FOURNISSEUR = 7,
    DOSSIER = 8,
    DEVIS = 9,
    AGENCE = 10,
    SOUS_TRAITANT = 11,
    BONCOMMANDE = 12
}